rust: add the langcomp benchmark test case
This is where rustc -O is 100x faster and serves to showcase
potential perf gains.

However, there was a little problem with type inference for

total += array[i]

where lhs is f64 and rhs is i32, needing a cast.

+++ tests/expected/binit.kt
@@ -7,7 +7,7 @@ fun bisect_right(data_: Array<Int>, item: Int): Int {
         if (item < data_[middle]) {
             high = middle
         } else {
-            low = (middle.toInt() + 1)
+            low = (middle + 1)
         }
     }
     return low
